+++ src/components/ItemResults.js
@@ -1,81 +1,28 @@
-import React, { useState } from "react";
+import React from "react";
 import MenuItemCard from "./MenuItemCard";
 import { Header, Divider, Container } from "semantic-ui-react";
 import RestaurantCard from "./RestaurantCard";
-import axios from "axios";
-
-
-
-// let defaultRestaurants
-// const handleSubmit = () => {
-//     // try {
-//         //     e.preventDefault()
-//         const result = axios.get('api/restaurants')
-//         .then(response => {
-//             defaultRestaurants = []
-//             const restDefault = response.data
-//             restDefault.map((restaurant) => {
-//                 defaultRestaurants.push(restaurant.name)
-//             })
-//             // console.log('axios')
-//             // console.log(defaultRestaurants)
-//             return (defaultRestaurants)
-//         })
-//         console.log('after axios')
-//         console.log(defaultRestaurants)
-//         return defaultRestaurants
-//         // } catch (error) {
-//             //     console.error(error.message)
-//             //     }
-// }
-
 
 function ItemResults ({ data }) {
-
-    // const [restaurant, setRestaurant] = useState()
-
-    const defaultRestaurants = []
-
-    const handleSubmit = () => {
-        axios.get('api/restaurants')
-        .then(response => {
-            const restDefault = response.data
-            restDefault.map((restaurant) => {
-                defaultRestaurants.push(restaurant)
-            })
-            // console.log('axios')
-        })
-        // setRestaurant(defaultRestaurants)
-        // console.log('after axios')
-        // console.log(defaultRestaurants)
-        return defaultRestaurants
-}
-
-// console.log(data)
-
-return (
-    <>
+    return (
+        <>
         {!data 
-        ? (         
-            <p>NO DATA</p>
-        )
-            : (
-                handleSubmit(),
-                console.log('data passed in return'),
-                // console.log(defaultRestaurants),
-                defaultRestaurants.map((name) => {
-                    console.log(name.name)
-                }),
-            // console.log( {data} ),
+        ? (
+            <p></p>)
+        : (
             <Container>
-            <br />hi
-            {defaultRestaurants.map((restaurant) => {
-                console.log('yes restaurant')
+            <br />
+            {data.map((item) => {
                 return (
-                    <>                  
+                    <>                   
                     <RestaurantCard 
-                        name={restaurant.name}
+                        name={item.restaurant.name}
                         />
+                    <MenuItemCard
+                        item={item.item}
+                        price={item.price}
+                        description={item.description}
+                        dietary={item.dietary}/>
                     </>
                 )
             }
